refactor(kftray-kube): promote shared_client to crate root
