+++ crates/kftray-kube/src/kube/proxy/lifecycle.rs
@@ -48,7 +48,7 @@ use rand::distr::{
     SampleString,
 };
 
-use crate::kube::shared_client::ServiceClientKey;
+use crate::shared_client::ServiceClientKey;
 use crate::port_forward_error::PortForwardError;
 use crate::registry::PORT_FORWARD_REGISTRY;
