+++ crates/kftray-kube/src/kube/service.rs
@@ -16,7 +16,7 @@ use crate::kube::operations::{
     get_services_with_annotation,
     list_all_namespaces,
 };
-use crate::kube::shared_client::ServiceClientKey;
+use crate::shared_client::ServiceClientKey;
 use crate::port_forward_error::PortForwardError;
 use crate::registry::PORT_FORWARD_REGISTRY;
 pub async fn retrieve_service_configs(
