+++ crates/kftray-kube/src/shared_client.rs
